Added nickname colors, changed font size, and more
This an attempt to clean up the way the chat looks a bit. Obviously this
is all subjective, so anything you guys have feedback on or don't like,
just let me know.

Changed the color of the input bar on joined from bright red to a cooler
(in the color sense) red.

Slightly reduced the top/bottom padding on chat messages.

Slightly reduced the font size of chat messages.

Changed the chat background color to a dark gray, and the text color to a
super light gray.

Added color to the nickname associated with any chat message. If it is
a regular chat message, nickname is blue; on a join, nickname is red; on
a button press, nickname is green.